Refactor GlobalStyle
