+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { BrowserRouter, Redirect, Route, Switch } from "react-router-dom";
 import * as P from "./Page/index";
-import GlobalStyle from "./Styles/GlobalStyles";
+import GlobalStyle from "./GlobalStyle/GlobalStyles";
 
 const App: React.FC = () => (
   <BrowserRouter>
